Set up logging earlier in the boot process
This fixes an issue where the ActionPack loggers do not write to STDOUT.

To reproduce:
    class TestController < ApplicationController
      def index
        Rails.logger.info "Logging with Rails.logger" # Printed to STDOUT
        logger.info "Logging with logger"             # Not printed to STDOUT
      end
    end

The ActionPack classes are set up by Railties initializers, including
setting the logger to Rails.logger. We need to set up the STDOUT logging
before the Railties initializers run, otherwise the loggers will be set
to the default Rails logger.