Include Rails 4 logger silence module when available
In Rails 3 the `#silence` method is available because Rails was monkey
patching stdlib's `Logger`, but in Rails 4 `ActiveSupport::Logger` was
introduced with a new module `LoggerSilence`, which provides it.

Since then the heroku logger was not having this method available due to
the use of stdlib's logger to log to `STDOUT`. However another `#silence`
method is added to `Kernel` by Rails, to capture a given stream (eg.
`stdout`), so calling `Rails.logger.silence` was actually resulting in
an exception:

    #<ArgumentError: wrong number of arguments (0 for 1)>
    ...gems/activesupport-4.2.1/lib/active_support/core_ext/kernel/reporting.rb:88:in `capture'`

Because the `capture/silence` method from `Kernel` expects the stream
argument.

This fixes it by creating a custom logger for heroku, which just
inherits from stdlib's `Logger` and includes Rails' `LoggerSilence` in
case it is defined.

Closes #14.